+++ AlphaFS/Filesystem/Directory Class/Directory.Encryption.cs
@@ -187,11 +187,16 @@ namespace Alphaleonis.Win32.Filesystem
       {
          string pathLp = Path.GetExtendedLengthPathInternal(null, path, pathFormat, GetFullPathOptions.RemoveTrailingDirectorySeparator | GetFullPathOptions.FullCheck);
 
-         var options = DirectoryEnumerationOptions.FilesAndFolders | DirectoryEnumerationOptions.AsLongPath | (recursive ? DirectoryEnumerationOptions.Recursive : 0);
+         var options = DirectoryEnumerationOptions.FilesAndFolders | DirectoryEnumerationOptions.AsLongPath;
 
-         // Process folders and files.
-         foreach (string fso in EnumerateFileSystemEntryInfosInternal<string>(null, pathLp, Path.WildcardStarMatchAll, options, PathFormat.LongFullPath))
-            File.EncryptDecryptFileInternal(true, fso, encrypt, PathFormat.LongFullPath);
+         // Process folders and files when recursive. 
+         if (recursive)
+         {
+            options |= DirectoryEnumerationOptions.Recursive;
+
+            foreach (string fso in EnumerateFileSystemEntryInfosInternal<string>(null, pathLp, Path.WildcardStarMatchAll, options, PathFormat.LongFullPath))
+               File.EncryptDecryptFileInternal(true, fso, encrypt, PathFormat.LongFullPath);
+         }
 
          // Process the root folder, the given path.
          File.EncryptDecryptFileInternal(true, pathLp, encrypt, PathFormat.LongFullPath);